detector/requirements.txt
pyyaml>=6.0
aiofiles>=23.0
aiohttp>=3.9
psutil>=5.9
python-dateutil>=2.8
